requirements.txt
crewai
crewai[tools]
openai==1.23.2
pinecone-client==3.2.2
python-dotenv==1.0.0
uuid
pathlib
uvicorn
fastapi
slowapi
httpx
jinja2
tiktoken
bs4
apscheduler
asyncio
langchain
langchain_experimental 
langchain_openai
langchain-groq
llama-parse
markdown2